Accumumator with epoch created.
